Revert "rk: ion: fix compilation error on arm64"
This reverts commit 6ba08d968d0938cac5d6d6d3ea0e71fd4aa05c27.

+++ include/linux/rockchip_ion.h
@@ -75,7 +75,7 @@ struct ion_phys_data {
 
 struct ion_share_id_data {
 	int fd;
-	unsigned long id;
+	unsigned int id;
 };
 
 #define ION_IOC_ROCKCHIP_MAGIC 'R'
